Remove bayes-opt + pass extra kwargs to self.model.fit()
