+++ requirements.txt
@@ -4,7 +4,6 @@ pandas>=1.1.5
 scikit-learn>=0.24.1
 xgboost>=1.4.2
 shap==0.38.1
-bayesian-optimization==1.2.0
 pyarrow>=5.0.0
 ipython>=7.16.1
 jedi==0.17.2
